Initial Pull
The background keeper is able to continuiously place an image as the
background if it detects a change. Currently this is all it does. I
would like to make sure that it only does this operation once in future
development. As of now, im happy with it as it will get done what i want
it to do.